Add files via upload
WIP of the Bar scene.  The models in this file are deliberately merged together into a single mesh to make selection easier.  Unity may disagree, but - for example - the tables are individual meshes on my end, so I can provide them as such if need be.  This file does not yet include UV mapping nor textures.

Steven( SamusDrake ).